Move WeakRefVirtualProxy to separate file
