+++ MyEssentialFeed/MyEssentialFeediOS/Feed UI/Composers/WeakRefVirtualProxy.swift
@@ -1,0 +1,29 @@
+//
+//  WeakRefVirtualProxy.swift
+//  MyEssentialFeediOS
+//
+//  Created by Tiago Chaves on 19/02/24.
+//
+
+import UIKit
+import MyEssentialFeed
+
+final class WeakRefVirtualProxy<T: AnyObject> {
+    private weak var object: T?
+    
+    init(_ object: T) {
+        self.object = object
+    }
+}
+
+extension WeakRefVirtualProxy: FeedLoadingView where T: FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel) {
+        object?.display(viewModel)
+    }
+}
+
+extension WeakRefVirtualProxy: FeedImageView where T: FeedImageView, T.Image == UIImage {
+    func display(_ model: FeedImageViewModel<UIImage>) {
+        object?.display(model)
+    }
+}
